pyproject.toml
[tool.poetry]
name = "proxy-request"
version = "0.1.0"
description = ""
authors = ["jack"]
readme = "README.md"
packages = [{include = "proxy_request"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
psycopg2 = "^2.9.6"
gunicorn = "^20.1.0"
django-filter = "^23.1"
requests = "^2.28.2"


[[tool.poetry.source]]
name = "gitlab"
url = "https://gitlab.com/api/v4/projects/41954132/packages/pypi/simple"
default = false
secondary = true


[tool.poetry.group.dev.dependencies]
django-cmds = {version = "^0.1.0", source = "gitlab"}
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
